LayoutTests/fast/forms/access-key-for-all-elements.html

<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<p>This test checks to see if accesskey attributes works all elements.</p>
<div id="test"></div>
<div id="console"></div>

<script>
if (window.layoutTestController)
    layoutTestController.dumpAsText();

var test = new Array("a","abbr","acronym","address","applet","area","article","aside","audio","b","base","basefont","bdo","bgsound","big","blockquote",
                     "body","br","button","canvas","caption","center","cite","code","col","colgroup","command","datalist","dd","del","details","dfn","dir","div","dl","dt",
                     "em","embed","fieldset","figcaption","figure","font","footer","form","frame","frameset","h1","h2","h3","h4","h5","h6","head","header","hgroup","hr","html",
                     "i","iframe","img","input","ins","isindex","kbd","keygen","label","layer","legend","li","link","listing","map","mark","marquee","menu","meta","meter","nav",
                     "nobr","noembed","noframes","nolayer","noscript","object","ol","optgroup","option","output","p","param","plaintext","pre","progress","q","rp","rt","ruby","s",
                     "samp","script","section","select","small","source","span","strike","strong","style","sub","summary","sup","table","tbody","td","textarea","tfoot","th","thead",
                     "title","tr","track","tt","u","ul","var","video","wbr","xmp");

for (var i =0 ; i < test.length; i++) {
    var parent = document.createElement(test[i]);
    document.body.appendChild(parent);
    var test_me = document.getElementsByTagName(test[i])[0];
    debug('Check for ' + test_me.tagName + ' tag');
    shouldBeDefined('test_me.accessKey');
    test_me.accessKey ='k';
    accesskey = test_me.accessKey; 
    shouldBe('accesskey',"'k'");
    debug('');
}
</script>

</body>
</html>
