LayoutTests/fast/forms/access-key-for-all-elements-expected.txt

This test checks to see if accesskey attributes works all elements.

Check for A tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for ABBR tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for ACRONYM tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for ADDRESS tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for APPLET tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for AREA tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for ARTICLE tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for ASIDE tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for AUDIO tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for B tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for BASE tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for BASEFONT tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for BDO tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for BGSOUND tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for BIG tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for BLOCKQUOTE tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for BODY tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for BR tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for BUTTON tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for CANVAS tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for CAPTION tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for CENTER tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for CITE tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for CODE tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for COL tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for COLGROUP tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for COMMAND tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for DATALIST tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for DD tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for DEL tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for DETAILS tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for DFN tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for DIR tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for DIV tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for DL tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for DT tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for EM tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for EMBED tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for FIELDSET tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for FIGCAPTION tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for FIGURE tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for FONT tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for FOOTER tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for FORM tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for FRAME tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for FRAMESET tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for H1 tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for H2 tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for H3 tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for H4 tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for H5 tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for H6 tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for HEAD tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for HEADER tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for HGROUP tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for HR tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for HTML tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for I tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for IFRAME tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for IMG tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for INPUT tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for INS tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for ISINDEX tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for KBD tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for KEYGEN tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for LABEL tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for LAYER tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for LEGEND tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for LI tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for LINK tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for LISTING tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for MAP tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for MARK tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for MARQUEE tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for MENU tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for META tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for METER tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for NAV tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for NOBR tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for NOEMBED tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for NOFRAMES tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for NOLAYER tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for NOSCRIPT tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for OBJECT tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for OL tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for OPTGROUP tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for OPTION tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for OUTPUT tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for P tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for PARAM tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for PLAINTEXT tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for PRE tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for PROGRESS tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for Q tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for RP tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for RT tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for RUBY tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for S tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for SAMP tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for SCRIPT tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for SECTION tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for SELECT tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for SMALL tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for SOURCE tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for SPAN tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for STRIKE tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for STRONG tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for STYLE tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for SUB tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for SUMMARY tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for SUP tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for TABLE tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for TBODY tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for TD tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for TEXTAREA tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for TFOOT tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for TH tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for THEAD tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for TITLE tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for TR tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for TRACK tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for TT tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for U tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for UL tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for VAR tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for VIDEO tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for WBR tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'

Check for XMP tag
PASS test_me.accessKey is defined.
PASS accesskey is 'k'
